Embed node-link component directly into controls
