+++ src/components/Controls.vue
@@ -1,4 +1,10 @@
 <script>
+import NodeLink from '@/components/NodeLink/NodeLink.vue';
+
+import { setUpProvenance } from "@/lib/provenance";
+import { getUrlVars } from "@/lib/utils";
+import { loadData } from "@/lib/multinet";
+
 /**
  * Demo Controls.  This component is meant to demo
  * the capabilities of the NodeLink vis.
@@ -8,50 +14,107 @@
  * implement the controls yourself.
  */
 export default {
-  props: {
-    nodeMarkerSize: {
-      type: Number,
-      default: 50
-    },
+  components: {
+    NodeLink,
+  },
 
-    nodeFontSize: {
-      type: Number,
-      default: 14,
-    },
-  }
+  data() {
+    /**
+     * State shared between the view (NodeLink graph)
+     * and the controller (dialog box or other UI).
+     */
+    return {
+      app: null,
+      provenance: null,
+      graphStructure: {
+        nodes: [],
+        links: []
+      },
+      nodeMarkerSize: 50,
+      nodeFontSize: 14,
+      workspace: null,
+      graph: null,
+      simOn: false,
+    };
+  },
+
+  /**
+   * This is the "entrypoint" into this application.
+   * Notice the v-if dependency in the template above.
+   */
+  async mounted() {
+    const { workspace, graph } = getUrlVars();
+    if (!workspace || !graph) {
+      throw new Error(
+        `Workspace and graph must be set! workspace=${workspace} graph=${graph}`
+      );
+    }
+    this.graphStructure = await loadData(workspace, graph);
+    const { provenance, app } = setUpProvenance(this.graphStructure.nodes);
+    this.app = app;
+    this.provenance = provenance;
+    this.workspace = workspace;
+    this.graph = graph;
+  },
 };
 </script>
 
 <template>
-  <v-card>
-    <v-card-title class="pb-6">MultiNet Node Link Controls</v-card-title>
+  <v-container>
+    <v-row class="flex-nowrap">
+      <!-- control panel content -->
+      <v-col class="shrink mt-4">
+        <v-card>
+          <v-card-title class="pb-6">MultiNet Node Link Controls</v-card-title>
 
-    <v-card-text>
-      <v-card-subtitle class="pb-0">Marker Size</v-card-subtitle>
-      <v-slider
-        :value="nodeMarkerSize"
-        @input="$emit('update:nodeMarkerSize', $event)"
-        :max="100"
-        :min="10"
-        hide-details
-      />
-      <p>{{ nodeMarkerSize }}</p>
-    </v-card-text>
+          <v-card-text>
+            <v-card-subtitle class="pb-0">Marker Size</v-card-subtitle>
+            <v-slider
+              v-model="nodeMarkerSize"
+              :min="10"
+              :max="100"
+              hide-details
+              />
+              <p>{{ nodeMarkerSize }}</p>
+          </v-card-text>
 
-    <v-card-text>
-      <v-card-subtitle class="pb-0">Font Size</v-card-subtitle>
-      <v-slider
-        :value="nodeFontSize"
-        @input="$emit('update:nodeFontSize', $event)"
-        :min="10"
-        :max="30"
-        hide-details
-      />
-      <p>{{ nodeFontSize }}</p>
-    </v-card-text>
+          <v-card-text>
+            <v-card-subtitle class="pb-0">Font Size</v-card-subtitle>
+            <v-slider
+              v-model="nodeFontSize"
+              :min="10"
+              :max="30"
+              hide-details
+              />
+              <p>{{ nodeFontSize }}</p>
+          </v-card-text>
 
-    <v-card-actions>
-      <v-btn small @click="$emit('restart-simulation')">Restart Simulation</v-btn>
-    </v-card-actions>
-  </v-card>
+          <v-card-actions>
+            <v-btn small @click="simOn = true">Restart Simulation</v-btn>
+          </v-card-actions>
+        </v-card>
+
+      </v-col>
+
+      <!-- node-link component -->
+      <v-col>
+        <v-row row wrap class="ma-0 pa-0">
+          <node-link
+            ref="nodelink"
+            v-if="workspace"
+            v-bind="{
+              graphStructure,
+              provenance,
+              app,
+              nodeMarkerHeight: nodeMarkerSize,
+              nodeMarkerLength: nodeMarkerSize,
+              nodeFontSize,
+              simOn,
+            }"
+            @restart-simulation="hello()"
+            />
+        </v-row>
+      </v-col>
+    </v-row>
+  </v-container>
 </template>
