+++ src/components/NodeLink/NodeLink.vue
@@ -126,6 +126,7 @@ export default {
         nodeFontSize,
         nodeMarkerLength,
         nodeMarkerHeight,
+        simOn,
         isDirected,
         isMultiEdge,
         attributes,
@@ -136,6 +137,7 @@ export default {
         nodeFontSize,
         nodeMarkerLength,
         nodeMarkerHeight,
+        simOn,
         isDirected,
         isMultiEdge,
         attributes,
